requirements.txt
Pillow==2.3.1
Wand==0.3.7
mercurial==2.9
wsgiref==0.1.2
